requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
groq==0.13.0
python-dotenv==1.0.1
